\name{imDefault}
\alias{imDefault}
%- Also NEED an `\alias' for EACH other topic documented here.
\title{Default region for an imagemap }
\description{
 Define a default region on an imagemap
}
\usage{
imDefault(...)
}
%- maybe also `usage' for other objects documented here.
\arguments{
  \item{\dots}{Extra arguments passed as name=value to the AREA tag}
}
\details{
 This function is used to specify a default region on an imagemap. This
 can be used to define a URL target for any clicks outside other
 specified regions.
}
\value{
 An object for adding to imagemaps.
}
\references{http://www.maths.lancs.ac.uk/Software/Imagemap/}
\author{Barry Rowlingson <B.Rowlingson@lancaster.ac.uk>}

\keyword{device}
