\name{imPoly}
\alias{imPoly}
%- Also NEED an `\alias' for EACH other topic documented here.
\title{Polygonal region on an imagemap}
\description{
 Specify a polygonal region on an imagemap
}
\usage{
imPoly(xy, ...)
}
%- maybe also `usage' for other objects documented here.
\arguments{
  \item{xy}{A 2-column matrix of coordinates}
  \item{\dots}{Extra arguments for the AREA tag}
}
\details{
 This function specifies a polygonal region for adding to an
 imagemap.
}
\value{
 An object for adding to imagemaps.
}
\references{http://www.maths.lancs.ac.uk/Software/Imagemap/}
\author{Barry Rowlingson <B.Rowlingson.ac.uk>}

\keyword{device}
