\name{toHTML}
\alias{toHTML}
%- Also NEED an `\alias' for EACH other topic documented here.
\title{Convert object to HTML}
\description{
 Convert an imagemap or a region to HTML
}
\usage{
toHTML(ob, im)
}
%- maybe also `usage' for other objects documented here.
\arguments{
  \item{ob}{an imagemap or region object }
  \item{im}{an imagemap object}
}
\value{
 Returns a character vector for the given object, one line per element.
}
\references{http://www.maths.lancs.ac.uk/Software/Imagemap/}
\author{Barry Rowlingson <B.Rowlingson@lancaster.ac.uk>}

\keyword{device}
